% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{calculate_accumulation}
\alias{calculate_accumulation}
\title{calculate_accumulation()}
\usage{
calculate_accumulation(temperature, precipitation, threshold_temperature)
}
\arguments{
\item{temperature}{Numeric. The temperature at the point of interest (in Celsius).}

\item{precipitation}{Numeric. The precipitation amount (in mm or other units).}

\item{threshold_temperature}{Numeric. The threshold temperature below which it snows (in Celsius).}
}
\value{
Numeric. The accumulation at the given point.
}
\description{
Calculate accumulation at a point
}
\examples{
calculate_accumulation(temperature = -2, precipitation = 10, threshold_temperature = 0)
}
