% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{calculate_lapsed_temperature}
\alias{calculate_lapsed_temperature}
\title{calculate_lapsed_temperature()}
\usage{
calculate_lapsed_temperature(elevation, elevation_station, temperature_station,
  lapse_rate)
}
\arguments{
\item{elevation}{Numeric. The elevation at the point of interest (in meters).}

\item{elevation_station}{Numeric. The elevation of the weather station (in meters).}

\item{temperature_station}{Numeric. The temperature at the weather station (in Celsius).}

\item{lapse_rate}{Numeric. The lapse rate (in Celsius per meter).}
}
\value{
Numeric. The lapsed temperature at the given elevation.
}
\description{
Calculate lapsed temperature
}
\examples{
calculate_lapsed_temperature(elevation = 1500, elevation_station = 1000,
temperature_station = 10, lapse_rate = -0.0065)
}
