% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{calculate_melt}
\alias{calculate_melt}
\title{calculate_melt()}
\usage{
calculate_melt(temperature, melt_factor)
}
\arguments{
\item{temperature}{average temperature}

\item{melt_factor}{a melting factor (m/day/degree).}
}
\description{
calculate the melting rate of the glacier (m/day)
}
\examples{
calculate_melt(temperature = 5, melt_factor = 0.1)
}
