% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{calculate_precip}
\alias{calculate_precip}
\title{calculate_precip()}
\usage{
calculate_precip(t)
}
\arguments{
\item{t}{time in days since beginning of simulation (currently deactivated)}
}
\description{
generate synthetic precipitation data (m/day)
}
\examples{
calculate_precip(t = 5)
}
