% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{glacier_net_balance_fn}
\alias{glacier_net_balance_fn}
\title{glacier_net_balance_fn()}
\usage{
glacier_net_balance_fn(zs, dt, Ts, Ps, melt_factor, T_threshold, lapse_rate)
}
\arguments{
\item{zs}{Array of elevations (with the weather station as datum).}

\item{dt}{The time step.}

\item{Ts}{Array of temperatures.}

\item{Ps}{Array of precipitations.}

\item{melt_factor}{The factor to compute melt amount.}

\item{T_threshold}{The temperature threshold for accumulation.}

\item{lapse_rate}{The lapse rate (temperature change per unit elevation change).}
}
\value{
A list containing the glacier net balance [m] and net balance at all points [m].
}
\description{
Calculate the glacier net balance and the net balance at each point.
}
