% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{net_balance_fn}
\alias{net_balance_fn}
\title{net_balance_fn()}
\usage{
net_balance_fn(dt, Ts, Ps, melt_factor, T_threshold)
}
\arguments{
\item{dt}{The time step.}

\item{Ts}{Array of temperatures.}

\item{Ps}{Array of precipitations.}

\item{melt_factor}{The factor to compute melt amount.}

\item{T_threshold}{The temperature threshold for accumulation.}
}
\value{
The net balance at a point.
}
\description{
Integrate the balance rate over time for given temperature and
             precipitation arrays to get the "net balance".
}
